Feature: fixing subtraction errors

+++ Calculator/subtraction.py
@@ -5,5 +5,5 @@ class Subtraction:
         pass
     @staticmethod
     def subtraction(a, b):
-        result = float(b) - float(a)
-        return result
+        result = int(b) - int(a)
+        return int(result)
